requirements.txt
pyyaml==6.0
tinkoff-investments==0.2.0b41
